Distinguish resource set from request complete in Glide's RequestCoordinators
A request might have it's thumbnail complete, but not it's full. It's
not complete, but it has a resource set. If it has a resource set, and a
parent's request fails, we shouldn't replace the thumbnail with the
error placeholder.

Complicating matters is that we have another way in which we want to
know if a resource is set. SingleRequest wants to know if its the first
request in the set so it can tell its RequestListener. That requires
walking up the request coordinator tree and then back down.

To solve the first issue, we check the requests explicitly and walk down
the tree to see if any has set a resource, regardless of its state.

To solve the second issue, we introduce a getRoot() method on
RequestCoordinator that can be used by SingleRequest to walk down the
entire tree.

FUTURE_COPYBARA_INTEGRATE_REVIEW=https://github.com/bumptech/glide/pull/3908 from sjudd:update_release_checks 59f5c1b3910857bbc7adf0c10215fc3c82d8c56c
PiperOrigin-RevId: 271661006